python/LLM/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
json_repair==0.49.0
lxml==6.0.0
openai==1.99.9
pillow==11.3.0
pydantic==2.11.7
pydantic_core==2.33.2
PyPDF2==3.0.1
python-docx==1.2.0
python-dotenv==1.1.1
regex==2025.7.34
reportlab==4.4.3
requests==2.32.4
sniffio==1.3.1
structlog==25.4.0
tiktoken==0.11.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
